// urban-management/features/map/types/admin-types.ts
import { OfficeProperties } from "@/features/office/types";
import { ReportStatus } from "@/features/report/types";

export type Geometry = {
  type: "Point";
  coordinates: [number, number]; // [lng, lat]
};

export type MapFilterRequest = {
  includeReports: boolean;
  includeOffices: boolean;
  statuses: ReportStatus[];
  categoryIds: string[];
  departmentIds: string[];
  keyword: string;
}

export type ReportProperties = {
  id: string;
  type: "report";
  title: string;
  category: string;
  aiConfidence: number;
  status: ReportStatus;
  description?: string;
  images?: string[];
  createdAt: string;
  updatedAt: string;
  address?: string;
};

export type FeatureProperties = ReportProperties | OfficeProperties;

export type Feature = {
  type: "Feature";
  properties: FeatureProperties;
  geometry: Geometry;
};

export type FeatureCollection = {
  type: "FeatureCollection";
  features: Feature[];
};